Restructure review routes using express router
